refactor ssg
