+++ packages/docusaurus-bundler/src/index.ts
@@ -15,6 +15,10 @@ export {
 } from './currentBundler';
 
 export {getMinimizers} from './minification';
-export {getHtmlMinifier, type HtmlMinifier} from './minifyHtml';
+export {
+  getHtmlMinifier,
+  type HtmlMinifier,
+  type HtmlMinifierType,
+} from './minifyHtml';
 export {createJsLoaderFactory} from './loaders/jsLoader';
 export {createStyleLoadersFactory} from './loaders/styleLoader';
